+++ mysql_perf.py
@@ -372,21 +372,6 @@ def run_program(args_array, func_dict, **kwargs):
         mysql_libs.disconnect(server)
 
 
-def is_pos_int(num):
-
-    """Function:  is_pos_int
-
-    Description:  Checks to see if number is an integer and positive.
-
-    Arguments:
-        (input) num -> Integer value.
-        (output) True|False -> Number is an integer and positive.
-
-    """
-
-    return isinstance(num, int) and num > 0
-
-
 def main():
 
     """Function:  main