+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - is_pos_int:  Checks to see if number is an integer and positive.
 
 ### Changed
+- mysql_stat_run:  Process status return from mongo_libs.ins_doc call.
 - run_program:  Replaced cmds_gen.disconnect with mysql_libs.disconnect call.
 - Removed unnesscary \*\*kwargs from function parameters lists.
 - config/mongo.py.TEMPLATE:  Changed configuration entry name and added a number of configuration entries.