+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Documentation updates.
 
 ### Removed
+- Removed is_pos_int function.
 - Removed datetime module.