+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Documentation updates.
 
 ### Removed
+- Removed lib.cmds_gen module.
 - Removed is_pos_int function.
 - Removed datetime module.